requirements.txt
fastapi
uvicorn
openai
python-dotenv
markdown2
langgraph
langchain